Separate pages and put in component folder
