+++ components/Map.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import {View, Text, StyleSheet, Button} from 'react-native';
+
+function MapScreen({navigation}) {
+  return (
+    <View style={{flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+      <Text>Map Screen!</Text>
+      <Button title="Go to Home" onPress={() => navigation.navigate('Home')} />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  welcomeText: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+});
+
+export default MapScreen;
